Move plugin card to components folder
